Fix CI issue caused by new cli package.
Recent versions of the cli package have added a new `vec-sep2` style
option for `cli_vec`, in an effort to improve their handling of Oxford
commas. This means `vec-last` is now ignored for any list of length two,
and the value of `vec-sep2` is used instead.

We only used that feature in one place, with "or" as the value for
`vec-last`. The new cli package broke a test by making that code print
the default `vec-sep2` value, i.e. "and", when there are only two
elements.

After looking at the context of that message however, I think "and"
actually makes more sense than "or". It is printing a list of paths that
haven't been found. We would want all of the paths in the list to be
present, not just one of them. The simplest fix is therefore to remove
the call to `cli_vec` and let cli use the default settings.

Finally, remove a call to `squote` on the missing paths and use a
`.path` tag in the message. The end result is the same.