Don't add a trailing newline to metadata files.
By default, the base writeLines function adds a trailing newline to the
end of the file, even when the input data did not contain any. While
this is generally an acceptable behaviour, in the case of metadata
files, we also hash the data and embed it in location files. Because the
hashing happens over the in-memory data, it does not include the
trailing newline. The hash of the on-disk file therefore ends up being
different from the hash stored in the location file.

It hasn't caused any practical issues because we trim leading and
trailing whitespaces when reading the files back, getting the original
serialized metadata that had been hashed. Similarly, when pushing a
metadata file to a remote server, the client trims whitespaces from the
metadata file. The behaviour is nevertheless confusing and it would be
better to have consitent contents and hashes all around.

The old behaviour of trimming whitespaces on read is preserved, as we
must maintain compatibility with old files that do contain the newline.