Allow orderly_shared_resource to be used with unnamed arguments.
The previous behaviour was inconsistent with orderly_resource and often
redundant, as users commonly want to copy the resource using the same
name as the original file.

It is now possible to use `orderly_shared_resource("foo.txt")`, in which
case the source and destination name are assumed to be the same.

The old syntax is still available in case the user wants to change the
name while copying, and it is even possible to mix-and-match, renaming
some files but not others.