Use vendored libgit2 in outpack_server on CI.
For some reason the CI has started failing on Windows because of libgit2
linking errors. Using a vendored libgit2 fixes the problem.

Normally libgit2-sys' build scripts should be clever ever to vendor it
when needed automatically, but for some reason that's not working and I
don't have the patience to figure out why right now. Using the vendored
version everywhere is the simplest fix.