Make files writable when they are exported.
Since [orderly2#141], orderly marks all files in the archive as
read-only. Similarly, files in the file store have always been
read-only. This helps prevent accidental corruption by users.

However this has had the undesirable side-effect of copying the files
with the read-only bit set, even when exporting for interactive use by
the user.

One scenario where this has come up is a user running
`orderly_dependency` interactively twice. They would expect the second
call to succeed and overwrite the existing files. However this was not
the case, and instead the second call fails as the read-only files
cannot be overwritten. The user ends up having to manually delete these
files (or mark them writable) everytime they want to update their copy.

This change marks all files that are exported as writable. This does not
apply to calls to `orderly_dependency` from a job context, in which case
the files are still read-only.

[orderly2#141]: https://github.com/mrc-ide/orderly2/pull/141