Replace uses of `fs::file_copy` with `file.copy`.
Unfortunately `fs::file_copy` has issues when working with read-only
files on a Samba mounted file system. `file.copy` doesn't have the same
problem. See mrc-5557 for more details.

This was tested manually and confirmed to work by running a basic report
with an `orderly_dependency` call on a Samba mount. Unfortunately it is
not something we can easily test for automatically.