Implement OAuth device flow for GitHub logins.
GitHub-based logins to Packit rely on users specifying a personal access
token when adding the location. This has a number of usability and
security disadvantages.

Users need to manually take steps in order to create a PAT, and they
need to make sure they have selected the right scope. It is quite likely
users would be tempted to create tokens with way more scopes than is
necessary.

The configured PAT would be stored in the orderly configuration, which
many users share and keep on network drives that are accessible to
everyone.

By using an OAuth flow to generate our own tokens, we remove the extra
steps the user needed to take and we can make sure it has just the right
scopes needed. Additionally the token cache implemented by httr2 is
located outside of the orderly repository, in the user's home directory.

The device flow works by showing the user an 8 character code in the
console, and a link to https://github.com/login/device. The user needs
to type in the code and approve the login. Meanwhile, the orderly2 client
polls the GitHub API until the authentication has been approved.

There are other OAuth flows we could have used, in particular the
authorization code flow (with PKCE). In this case, a link is printed to
the console and the user needs to follow it and approve the login. They
are then redirected to a localhost server, which has been started by
orderly2. This removes the need to need with the 8 character code of the
device flow, but it is a bit finicky and assumes orderly2 is running on
the same machine as the user's browser, which may not be the case when
runing over SSH. For this reason, the device flow is preferred.

It is still possible to specify a PAT as before, in which case the OAuth
flow is skipped. This can be useful in non-interactive situations, such
as a CI pipeline.

Note that there are still fundamental issues with this model. In
particular, the Github tokens are not bound to the application. This
means that a malicious application could manage to get a user to sign in
to it, obtain a access token for the user, and use this to login to
Packit. From just a GitHub `read:org` scope, the app is able to escalate
to full impersonation of the user on the Packit server.

Additionally it still tightly couples the orderly2 client with
GitHub as an authentication method, prohibiting the use of other
authentication methods, such as basic username / password.

Ideally we would remove the use of Github Access tokens as a way to
login to Packit. Instead, Packit should itself implement an OAuth2
Authorization Server. orderly2 would perform the device flow directly
against Packit, and directly retrieve a Packit JWT in exchange. During
the device flow, when the user is directed to the Packit website to
enter the user code, Packit may require the user to login. When this
happens, a new, parallel OAuth flow would take place, this time with
GitHub as the OAuth AS and Packit as the OAuth client (using the
authorization code flow). When the nested flow is complete and login is
succesful, the user would be prompted to type in the code and complete
the flow initiated by orderly2.

With this approach, orderly2 has no involvement with GitHub at all, and
Packit can use another authentication method (or even no authentication).
This would be completely transparent to the orderly2 client. It also
provides tighter binding between GitHub and Packit, as Packit fetches
the access token directly from the former, and cannot be supplied a
token from an arbitrary application.