Don't automatically unbox vectors when making requests.
When upgrading to httr2, I rewrote some of the http client code and
forgot to keep the `auto_unbox = FALSE` flag. This has caused uploads to
outpack_server to break, as requests were now malformed.

The issue is that it is ambiguous whether `list(x = 1)` should be
serialized as `{"x": 1}` or as `{"x": [ 1 ]}`, because R does not
differentiate between scalar and vectors of length one.

By default, httr2 applies `auto_unbox = TRUE`, which leads to all length
1 vectors being unboxed into a JSON scalar. By explictly choosing
setting that flag to `FALSE`, we may override this behaviour and keep
all vectors as vectors by default. Scalars need to be marked as such
with the `scalar(x)` function (defined as `jsonlite::unbox(x)`).

Frustratingly this did not get caught by our testsuite because we were
always pushing more than one packet at a time. I've added a test for
this particular case, pushing one packet and one file, to make sure
regressions don't get introduced in the future.