Use the fs package to copy files.
It has more bells and whistles, and is often more performant. In
particular, it makes use of `CopyFileW` (on Windows) and
`copy_file_range` (on Linux) when applicable. This offers a big speedup
when copying a file within a network drive, which is a common use case
for orderly.

The reason why `file.copy` was introduced in #165 is that copying
read-only files with the `fs` package wasn't working properly on Samba
drives. It turns out this situation only occurs when `overwrite = TRUE`.
When the flag is set to FALSE, the copy succeeds. To work around this we
do an initial check before the copy to delete any files in the
destination path, and copy with `overwrite = FALSE`. This was manually
tested by running the test suite with the `TMPDIR` environment variable
set to a path on a network drive.