Include the file path when a JSON decoding error happens.
If the metadata store is corrupted for some reason, jsonlite throws an
error with details of the syntax error, but does not include any mention
of the file path.

This adds a wrapper around `jsonlite::parse_json` which catches the
error and includes the file path. It still preserves the original error
message, using [{rlang} error chaining][error-chain].

[error-chain]: https://rlang.r-lib.org/reference/topic-error-chaining.html