Add support for specifying Packit tokens directly.
As part of experimenting with new authentication methods for Packit, it
is useful to be able to provide Packit tokens directly to orderly2.
Until now, orderly2 would only accept GitHub tokens, and would perform a
request to exchange it for a Packit token.

Thankfully, GitHub tokens have a [well defined and documented
prefix][ghblog], of the form `ghp_`, `gho_`, ... The underscore makes
sure it so that it never could never be confused with a JWT. Thanks to
this, the distinction between Packit and GitHub tokens is unambiguous.

Eventually we may want to implement some of these authentication methods
in orderly2 itself, but it the meantime this makes experimentation
easier.

[ghblog]: https://github.blog/engineering/platform-security/behind-githubs-new-authentication-token-formats/