Support `environment:` from an interactive session.
The code path for interactive sessions, through orderly_copy_files, was
not propagating the environment from the caller to `orderly_search`. The
broke support for `environment:`-style queries, but only when running
interactively.

In orderly_copy_files (and orderly_dependency), the `envir` parameter
actually has two purposes: it provides values for `environment:` and it
provides values that are substituted in the file paths to be copied.
This dual usage causes a bit of confusion and may be what led to this
issue. We may want to tidy up some of this in the future, possibly with
the help of rlang's quosures and quasiquotation support.