Perform device flow authentication against Packit.
In the past orderly2 would perform device flow authentication against
GitHub, retrieve a GitHub token and use that to trade it against a
Packit token.

Packit now supports device flow authentication natively. Orderly2 can
directly get a Packit token, without having to trade anything. If
necessary, the device flow may involve logging in to Packit via GitHub,
but that happens in the user's browser and is completely transparent
from the user's perspective.

It also means the authentication is completely standard and works with
the off-the-shelf components from httr2. We do not have to think about
caching the Packit token anymore, as that is now done automatically by
httr2.

We also probably do not need to worry about expiry and renewals, which
is something that we had never handled properly (the tokens were only
cached in memory, so restarting the R session regularly avoided this
problem). httr2 will automatically initiate authentication when a new
token is needed.

The old authentication method has been removed from orderly2 completely.
We will wait until all Packit instances support device-flow before
merging this change.