Allow orderly_dependency to be used with an empty list of files.
This is definitely a niche use case, but I don't see a good reason not
to support it. For the most part, everything worked already.

The only issue was the `ragged_replace` function which returned NULL
when the output was empty instead of a vector of the original type. This
would cause columns to be missing from data frames (`dataframe(x =
NULL)` has no columns, rather than an empty `x`), as well as cause
problems with `fs` functions (eg. `fs::file_exists(NULL)` is an error,
but `fs::file_exists(character(0))` works and returns an empty logical
vector).

Using the vctrs package makes it easy to write code that has consistent
and predictable types. httr2 already depends on that package, so it is
not actually a new dependency.