Fix the documentation of orderly_copy_files (and related).
The documentation was claiming that the `files` argument can be a
dataframe with here and there columns, but that is not actually the
case. Only character vectors (and lists for some reason) are actually
supported.

The function that processes the files list is `validate_file_from_to`
and has the following comment:

> Later, we can expand this to support a data.frame too perhaps?

Until then we shouldn't be advertising an inexistent feature.