Detect issues with interactive use of orderly_parameters.
When `orderly_parameters` is called, we actually ignore its arguments
and instead use the invocation of it found in the source file of the
report. We do this because we need to know the name of the variable the
parameters are assigned to, which cannot be done just from looking at
the arguments.

This can cause issues when running interactively in RStudio, running
line-by-line with Ctrl-Enter but without saving the file first. The user
may have added a parameter, not saved the file and run the
`orderly_parameters` line. orderly would use the list of parameters it
finds on disk instead of the ones in the text editor's buffer.

We could be really clever and use the RStudio API to read the unsaved
buffer instead, but that's a bit over the top and won't work with other
IDEs that have a similar interactive interface.

The easy solution is to compare the arguments given to
orderly_parameters against the ones that were read from the on-disk
file, and throw an error if these don't match.